requirements.txt
# Core LangChain and LangGraph
langchain>=0.2.0
langgraph>=0.2.0
langchain-core>=0.2.0

# MCP Integration
langchain-mcp-adapters>=0.1.0

# LLM Providers
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0

# Vector Database
qdrant-client>=1.8.0
langchain-qdrant>=0.1.0

# Evaluation
ragas>=0.1.0
datasets>=2.14.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
fastembed